+++ docs/changes.rst
@@ -97,6 +97,8 @@ Compatibility notes:
   - removed --save-space option (does not change behaviour)
 - removed --bypass-lock option
 - removed borg config command (only worked locally anyway)
+- compact command now requires access to the borg key if the repo is encrypted
+  or authenticated
 - using --list together with --progress is now disallowed (except with --log-json), #7219
 - the --glob-archives option was renamed to --match-archives (the short option
   name -a is unchanged) and extended to support different pattern styles: